fix(NumberParser): Change location of NumberParser
Change location of NumberParser
because it is only used in inputView package
