+++ src/test/java/lotto/util/NumberParserTest.java
@@ -3,6 +3,7 @@ package lotto.util;
 import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.List;
+import lotto.view.NumberParser;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
